general clean up, no * imports, rename Main file
