+++ src/main/kotlin/model/Dividends.kt
