+++ src/main/kotlin/util/Functions.kt
@@ -1,4 +1,4 @@
-package robinhood
+package util
 
 import io.ktor.client.statement.HttpResponse
 import io.ktor.utils.io.jvm.javaio.toInputStream
