+++ src/main/kotlin/util/UuidSerializer.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.encoding.Decoder
 import kotlinx.serialization.encoding.Encoder
 import java.time.OffsetDateTime
 import java.time.format.DateTimeFormatter
-import java.util.*
+import java.util.UUID
 
 
 object UUIDSerializer : KSerializer<UUID> {
